requirements.txt
fastapi[standard]
langchain-google-genai
langchain-chroma
langchain 
langchain-core
langchain-community
pypdf
streamlit
python-dotenv

